Rename endpoints package to constants
