+++ src/main/java/teammates/ui/constants/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains all constants to be shared with the front-end.
+ */
+package teammates.ui.constants;
