--- src/main/java/teammates/ui/endpoints/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Contains all API endpoints.
- */
-package teammates.ui.endpoints;
